moved logic to appropriate app
