--- streetteam/apps/twilio_integration/templates/phone_number.html
@@ -1,5 +1,0 @@
-<form action="/integration/confirm_phone_number/" method="post">
-  {% csrf_token %}
-  {{ form }}
-  <input type="submit" value="Submit">
-</form>
