+++ streetteam/apps/users/templates/enter_verification_code.html
@@ -1,4 +1,4 @@
-<form action="/integration/enter_verification_code/" method="post">
+<form action="{% url 'verify_code_send_via_sms' %}" method="post">
   {% csrf_token %}
   {{ form }}
   <input type="submit" value="Submit">
